requirements.txt
numpy==2.0.0
Pillow==10.4.0
